bug fixed

+++ DataGen.py
@@ -31,7 +31,7 @@ class Batch:
 
 def data_gen(V, batch, nbatches):
     for i in range(nbatches):
-        data = torch.from_numpy(np.random.randint(1, V, size=(batch, 10)))
+        data = torch.from_numpy(np.random.randint(1, V, size=(batch, 10))).to(torch.int64)
         data[:, 0] = 1
         src = Variable(data, requires_grad=False)
         tgt = Variable(data, requires_grad=False)
